Fix config migration code

+++ stroom-config/stroom-config-global-impl-db/src/main/java/stroom/config/global/impl/db/migration/V7_0_0_2__property_rename.java
@@ -21,7 +21,7 @@ import static stroom.config.impl.db.stroom.tables.Config.CONFIG;
 
 @SuppressWarnings("unused")
         // used by FlyWay
-class V7_0_0_2__property_rename implements JdbcMigration {
+public class V7_0_0_2__property_rename implements JdbcMigration {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(V7_0_0_2__property_rename.class);
 
@@ -30,7 +30,8 @@ class V7_0_0_2__property_rename implements JdbcMigration {
     static {
 
         // TODO what do we do about mapping c3po pool props to hikari?
-        FROM_TO_MAP.put("stroom.aboutHtml", "stroom.ui.aboutHTML");
+
+        FROM_TO_MAP.put("stroom.aboutHTML", "stroom.ui.aboutHtml");
         FROM_TO_MAP.put("stroom.activity.chooseOnStartup", "stroom.ui.activity.chooseOnStartup");
         FROM_TO_MAP.put("stroom.activity.editorBody", "stroom.ui.activity.editorBody");
         FROM_TO_MAP.put("stroom.activity.editorTitle", "stroom.ui.activity.editorTitle");
@@ -54,6 +55,7 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.contentPackImportEnabled", "stroom.contentPackImport.enabled");
         FROM_TO_MAP.put("stroom.dashboard.defaultMaxResults", "stroom.ui.defaultMaxResults");
         FROM_TO_MAP.put("stroom.databaseMultiInsertMaxBatchSize", "stroom.core.databaseMultiInsertMaxBatchSize");
+        // TODO how to map connection pool props?
         // stroom.db.connectionPool.acquireIncrement
         // stroom.db.connectionPool.acquireRetryAttempts
         // stroom.db.connectionPool.acquireRetryDelay
@@ -74,6 +76,7 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.fileSystemCleanDeleteOut", "stroom.data.store.fileSystemCleanDeleteOut");
         FROM_TO_MAP.put("stroom.fileSystemCleanOldAge", "stroom.data.store.fileSystemCleanOldAge");
         FROM_TO_MAP.put("stroom.helpUrl", "stroom.ui.helpUrl");
+
         // same names in 6 & master
         // stroom.index.ramBufferSizeMB
         // stroom.index.writer.cache.coreItems
@@ -81,6 +84,7 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         // stroom.index.writer.cache.minItems
         // stroom.index.writer.cache.timeToIdle
         // stroom.index.writer.cache.timeToLive
+
         FROM_TO_MAP.put("stroom.internalstatistics.benchmarkCluster.docRefs", "stroom.statistics.internal.benchmarkCluster");
         FROM_TO_MAP.put("stroom.internalstatistics.cpu.docRefs", "stroom.statistics.internal.cpu");
         FROM_TO_MAP.put("stroom.internalstatistics.eventsPerSecond.docRefs", "stroom.statistics.internal.eventsPerSecond");
@@ -98,13 +102,17 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.jdbcDriverUsername", "stroom.core.connection.jdbcDriverUsername");
         FROM_TO_MAP.put("stroom.jpaDialect", "stroom.core.hibernate.dialect");
         FROM_TO_MAP.put("stroom.jpaHbm2DdlAuto", "stroom.core.hibernate.jpaHbm2DdlAuto");
+
         // same names in 6 & master
         // stroom.lifecycle.enabled
         // stroom.lifecycle.executionInterval
+
         // Now defined in a doc entity
         // stroom.kafka.bootstrap.servers
+
         // Now handled by auth service
         // stroom.loginHTML
+
         FROM_TO_MAP.put("stroom.maintenance.message", "stroom.ui.maintenanceMessage");
         FROM_TO_MAP.put("stroom.maintenance.preventLogin", "stroom.security.authentication.preventLogin");
         FROM_TO_MAP.put("stroom.maxAggregation", "stroom.proxyAggregation.maxAggregation");
@@ -112,18 +120,22 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.maxStreamSize", "stroom.proxyAggregation.maxStreamSize");
         FROM_TO_MAP.put("stroom.namePattern", "stroom.ui.namePattern");
         FROM_TO_MAP.put("stroom.node", "stroom.node.node");
+
         // Same names in 6 & master
         // stroom.node.status.heapHistogram.classNameMatchRegex
         // stroom.node.status.heapHistogram.classNameReplacementRegex
         // stroom.node.status.heapHistogram.jMapExecutable
+
         FROM_TO_MAP.put("stroom.pageTitle", "?");
         // Same names in 6 & master
         // stroom.pipeline.appender.maxActiveDestinations
         // stroom.pipeline.xslt.maxElements
+
         // These are now all part of ProxyConfig that does not get written to the DB and thus has no prop name.
         // stroom.proxy.store.dir ?
         // stroom.proxy.store.format ?
         // stroom.proxy.store.rollCron ?
+
         FROM_TO_MAP.put("stroom.proxyDir", "stroom.proxyAggregation.proxyDir");
         FROM_TO_MAP.put("stroom.proxyThreads", "stroom.proxyAggregation.proxyThreads");
         FROM_TO_MAP.put("stroom.query.history.daysRetention", "stroom.queryHistory.daysRetention");
@@ -132,23 +144,28 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.query.infoPopup.title", "stroom.ui.query.infoPopup.title");
         FROM_TO_MAP.put("stroom.query.infoPopup.validationRegex", "stroom.ui.query.infoPopup.validationRegex");
         FROM_TO_MAP.put("stroom.rack", "stroom.node.rack");
+
         // Same name in 6 & master
         // stroom.search.extraction.maxThreads
         // stroom.search.extraction.maxThreadsPerTask
         // stroom.search.maxBooleanClauseCount
         // stroom.search.maxStoredDataQueueSize
+
         FROM_TO_MAP.put("stroom.search.process.defaultRecordLimit", "stroom.ui.process.defaultRecordLimit");
         FROM_TO_MAP.put("stroom.search.process.defaultTimeLimit", "stroom.ui.process.defaultTimeLimit");
+
         // Same name in 6 & master
         // stroom.search.shard.maxDocIdQueueSize
         // stroom.search.shard.maxThreads
         // stroom.search.shard.maxThreadsPerTask
         // stroom.search.storeSize
+
         FROM_TO_MAP.put("stroom.security.apitoken", "stroom.security.authentication.apiToken");
         FROM_TO_MAP.put("stroom.security.userNamePattern", "stroom.security.authentication.userNamePattern");
         FROM_TO_MAP.put("stroom.serviceDiscovery.curator.baseSleepTimeMs", "stroom.serviceDiscovery.curatorBaseSleepTimeMs");
         FROM_TO_MAP.put("stroom.serviceDiscovery.curator.maxRetries", "stroom.serviceDiscovery.curatorMaxRetries");
         FROM_TO_MAP.put("stroom.serviceDiscovery.curator.maxSleepTimeMs", "stroom.serviceDiscovery.curatorMaxSleepTimeMs");
+
         // TODO need to figure out what we are doing with service disco
         // stroom.serviceDiscovery.enabled
         // stroom.serviceDiscovery.servicesHostNameOrIpAddress
@@ -171,6 +188,7 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         // ? stroom.services.stroomStats.docRefType ?
         // ? stroom.services.stroomStats.name ?
         // ? stroom.services.stroomStats.version ?
+
         FROM_TO_MAP.put("stroom.services.stroomStats.internalStats.eventsPerMessage", "stroom.statistics.hbase.eventsPerMessage");
         FROM_TO_MAP.put("stroom.services.stroomStats.kafkaTopics.count", "stroom.statistics.hbase.kafkaTopics.count");
         FROM_TO_MAP.put("stroom.services.stroomStats.kafkaTopics.value", "stroom.statistics.hbase.kafkaTopics.value");
@@ -179,8 +197,10 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.splash.enabled", "stroom.ui.splash.enabled");
         FROM_TO_MAP.put("stroom.splash.title", "stroom.ui.splash.title");
         FROM_TO_MAP.put("stroom.splash.version", "stroom.ui.splash.version");
+
         // No longer used, stat doc entities are totally separate
         // stroom.statistics.common.statisticEngines
+
         // TODO what to do about
         // ? stroom.statistics.sql.db.connectionPool.acquireIncrement
         // ? stroom.statistics.sql.db.connectionPool.acquireRetryAttempts
@@ -195,26 +215,32 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         // ? stroom.statistics.sql.db.connectionPool.minPoolSize
         // ? stroom.statistics.sql.db.connectionPool.numHelperThreads
         // ? stroom.statistics.sql.db.connectionPool.unreturnedConnectionTimeout
+
         FROM_TO_MAP.put("stroom.statistics.sql.jdbcDriverClassName", "stroom.statistics.sql.connection.jdbcDriverClassName");
         FROM_TO_MAP.put("stroom.statistics.sql.jdbcDriverPassword", "stroom.statistics.sql.connection.jdbcDriverPassword");
         FROM_TO_MAP.put("stroom.statistics.sql.jdbcDriverUrl", "stroom.statistics.sql.connection.jdbcDriverUrl");
         FROM_TO_MAP.put("stroom.statistics.sql.jdbcDriverUsername", "stroom.statistics.sql.connection.jdbcDriverUsername");
+
         // Same names in 6 & master
         // stroom.statistics.sql.maxProcessingAge
         // stroom.statistics.sql.search.fetchSize
         // stroom.statistics.sql.search.maxResults
         // stroom.statistics.sql.search.resultHandlerBatchSize
         // stroom.statistics.sql.statisticAggregationBatchSize
+
         // TODO lots of options for these; stroom.data.store, stroom.policy, stroom.process
         // ? stroom.stream.deleteBatchSize ? stroom.data.store.deleteBatchSize
         // ? stroom.stream.deletePurgeAge ? stroom.data.store.deletePurgeAge
         // ? stroom.streamAttribute.deleteAge ? stroom.process.deleteAge
         // ? stroom.streamAttribute.deleteBatchSize ?
+
         FROM_TO_MAP.put("stroom.streamTask.assignTasks", "stroom.process.assignTasks");
         FROM_TO_MAP.put("stroom.streamTask.createTasks", "stroom.process.createTasks");
+
         // TODO lots of options for these; stroom.data.store, stroom.policy, stroom.process
         // ? stroom.streamTask.deleteAge ? stroom.process.deleteAge
         // ? stroom.streamTask.deleteBatchSize ?
+
         FROM_TO_MAP.put("stroom.streamTask.fillTaskQueue", "stroom.process.fillTaskQueue");
         FROM_TO_MAP.put("stroom.streamTask.queueSize", "stroom.process.queueSize");
         FROM_TO_MAP.put("stroom.streamstore.preferLocalVolumes", "stroom.volumes.preferLocalVolumes");
@@ -230,14 +256,16 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         FROM_TO_MAP.put("stroom.theme.tube.opacity", "stroom.ui.theme.tubeOpacity");
         FROM_TO_MAP.put("stroom.theme.tube.visible", "stroom.ui.theme.tubeVisible");
         FROM_TO_MAP.put("stroom.unknownClassification", "stroom.ui.theme.tubeVisible");
+
         // Same name in 6 & master
         // stroom.volumes.createDefaultOnStart
+
         FROM_TO_MAP.put("stroom.welcomeHTML", "stroom.ui.welcomeHtml");
     }
 
     private final ConfigMapper configMapper;
 
-    V7_0_0_2__property_rename() {
+    public V7_0_0_2__property_rename() {
 
         // initialise the config mapper so we can validate property keys
         configMapper = new ConfigMapper(new AppConfig());
@@ -248,13 +276,17 @@ class V7_0_0_2__property_rename implements JdbcMigration {
         try {
             final DSLContext create = DSL.using(connection, SQLDialect.MYSQL);
 
+            // This line should only be un-commented for manual testing in development
+//            loadTestDataForManualTesting(create);
+
+            // Rename some property names
             FROM_TO_MAP.entrySet().stream()
                     .sorted(Comparator.comparing(Map.Entry::getKey))
                     .forEach(entry -> {
                         final String from = entry.getKey();
                         final String to = entry.getValue();
 
-                        Optional<ConfigRecord> optRec = create
+                        final Optional<ConfigRecord> optRec = create
                                 .selectFrom(CONFIG)
                                 .where(CONFIG.NAME.eq(from))
                                 .fetchOptional();
@@ -272,20 +304,41 @@ class V7_0_0_2__property_rename implements JdbcMigration {
                         });
                     });
 
+            // Remove property entries that don't map to the object model
             create
                     .selectFrom(CONFIG)
                     .fetch()
                     .forEach(configRecord -> {
                         boolean isToPathValid = configMapper.validatePropertyPath(configRecord.getName());
-                        LOGGER.warn("Property {} with value [{}] is not valid a valid property " +
-                                "in the object model. It will be deleted.",
-                                configRecord.getName(), configRecord.getVal());
-                        configRecord.delete();
+                        if (!isToPathValid) {
+                            LOGGER.warn("Property {} with value [{}] is not valid a valid property " +
+                                            "in the object model. It will be deleted.",
+                                    configRecord.getName(), configRecord.getVal());
+                            configRecord.delete();
+                        }
                     });
 
         } catch (final RuntimeException e) {
             LOGGER.error("Error renaming property", e);
             throw e;
         }
+    }
+
+    private void loadTestDataForManualTesting(final DSLContext create) {
+
+        LOGGER.warn("Loading test data - Not for use in prod");
+        final Map<String, String> testDataMap = Map.of(
+
+                "stroom.aboutHTML", "myHtml",
+                "stroom.advertisedUrl", "myUrl",
+                "stroom.unknownProp", "some value"
+        );
+
+        testDataMap.forEach((key, value) ->
+                create
+                        .insertInto(CONFIG)
+                        .columns(CONFIG.NAME, CONFIG.VAL)
+                        .values(key, value)
+                        .execute());
     }
 }
